% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_browser_emojis.R
\docType{data}
\name{code_browser_emojis}
\alias{code_browser_emojis}
\title{code_browser_emojis}
\format{
3 columns 1853 rows
\describe{
\item{Message}{Data frame with emoji conversions}
}
}
\usage{
data("code_browser_emojis")
}
\description{
data frame with columns for emoji converting
}
\keyword{datasets}
