% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{data_param}
\alias{data_param}
\title{Helper for consistent documentation of \code{data}}
\arguments{
\item{data}{Name of your Data Frame or Tibble object}
}
\description{
Use \verb{@inheritParams data_param} to consistently document \code{data}.
}
\keyword{internal}
