% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{entities}
\alias{entities}
\title{entities}
\format{
4 columns 193 rows
\describe{
\item{token}{regex pattern for product}
\item{replacement}{string to replace with}
\item{n}{count of entity in original data}
\item{word_count}{number of words entity's string comprises}
}
}
\usage{
data("entities")
}
\description{
data frame with entities for converting and removing
}
\keyword{internal}
