% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_accents.R
\name{limpiar_accents}
\alias{limpiar_accents}
\title{Clean accented characters}
\usage{
limpiar_accents(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}
}
\value{
Data Frame or Tibble object with accents in the text variable replaced
}
\description{
Useful for reducing overall number of tokens.
Warning: removing accents results in loss of information, so should be done with care.
}
\examples{
limpiar_examples \%>\% dplyr::select(mention_content)
limpiar_examples \%>\% limpiar_accents() \%>\% dplyr::select(mention_content)


}
