% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_df.R
\name{limpiar_df}
\alias{limpiar_df}
\title{limpiar_df}
\usage{
limpiar_df(df, text_var)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{text_var}{The text variable/character vector}
}
\value{
The original Data Frame with the text variable cleaned
}
\description{
used on a data frame to clean the column names, remove duplicates, empty rows and retweets,
turn the text variable to all lower case (advisable to do any part of speech tagging before using).
Function also cleans accents, common urls and trims excess white spaces.
}
\examples{
print("hello world")
}
