% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_duplicates.R
\name{limpiar_duplicates}
\alias{limpiar_duplicates}
\title{limpiar_duplicates}
\usage{
limpiar_duplicates(data, text_var = .data$mention_content)
}
\arguments{
\item{data}{Data Frame or Tibble object}

\item{text_var}{Name of the text variable/character vector}
}
\value{
text variable/character with duplicates removed
}
\description{
Removes duplicate posts and 'deleted or protected mentions'
}
\examples{
df <- data.frame(text_variable = cbind(c("Deleted or protected mention", "hello", "goodbye", "goodbye")))
limpiar_duplicates(df, text_variable)
}
