% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_ex_subreddits.R
\name{limpiar_ex_subreddits}
\alias{limpiar_ex_subreddits}
\title{Quickly extract subreddits from a link variable}
\usage{
limpiar_ex_subreddits(df, url_var = permalink)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{url_var}{The variable containing URLS e.g. \code{permalink}}
}
\value{
df with an additional column
}
\description{
Quickly extract subreddits from a link variable
}
