% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_link_click.R
\name{limpiar_link_click}
\alias{limpiar_link_click}
\title{Prepare a URL column to be clickable in Shiny/Data Table}
\usage{
limpiar_link_click(df, url_var)
}
\arguments{
\item{df}{Data Farame or Tibble Object}

\item{url_var}{URL Column}
}
\value{
data frame with URL column edited to be clickable
}
\description{
Will allow you to click the hyperlink to load a URL, e.g. for selecting an image.
Make sure that DataTable is rendered with the argument 'escape = FALSE' or column will be all text.=
}
\examples{
\dontrun{
Example 1:
df \%>\% limpiar_link_click(permalink)

Example 2:
limpiar_link_click(data, mention_url)
}
}
