% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_link_click.R
\name{limpiar_link_click_reverse}
\alias{limpiar_link_click_reverse}
\title{Reverses (inverts) limpiar_link_click}
\usage{
limpiar_link_click_reverse(df, url_var)
}
\arguments{
\item{df}{Data Farame or Tibble Object}

\item{url_var}{URL Column}
}
\value{
Data frame with the url_var in original form
}
\description{
Undoes the effects of the limpiar_link_click function, giving you the original url variable back.
}
\examples{
df <- LimpiaR::limpiar_examples[1, ]

df <- df \%>\% limpiar_link_click(mention_url)
df$mention_url
df \%>\% limpiar_link_click_reverse(mention_url)

}
