% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_na_cols.R
\name{limpiar_na_cols}
\alias{limpiar_na_cols}
\title{Clean NA-heavy columns from a Data Frame or Tibble}
\usage{
limpiar_na_cols(df, threshold)
}
\arguments{
\item{df}{The Data Frame or Tibble object}

\item{threshold}{Threshold of non-NA entries a column must exceed to be retained.}
}
\value{
Data Frame or Tibble with NA-heavy columns purged
}
\description{
Remove columns whose proportion of NAs is higher than determined threshold.
Setting threshold of 0.25 asks R to remove all columns with 25\% or more NA values.
Can be useful when dealing with large data frames, where many columns are redundant.
}
\examples{
a <- cbind(c(NA, NA, NA, NA, NA))
b <- cbind(c(NA, NA, 2, NA, 3))
c <- cbind(c(1, 2, 3, 4, 5))
df <- tibble::as_tibble(data.frame(a, b, c))
df <- limpiar_na_cols(df, 0.4)

}
