% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pos_download_model.R
\name{limpiar_pos_download_model}
\alias{limpiar_pos_download_model}
\title{Download UDPipe models to begin Parts of Speech Analysis}
\usage{
limpiar_pos_download_model(language, model_directory, overwrite = TRUE)
}
\arguments{
\item{language}{The chosen language that the user wishes to select. There are 65 options to choose from}

\item{model_directory}{A string pointing to a specific directory informing the function where to save the model}

\item{overwrite}{Whether to overwrite any pre-exisiting models saved in this location. \code{overwrite} is set to \code{TRUE} by default, meaning it will download the model and overwrite the file if the file already existed. If set to \code{FALSE}, the model will only be downloaded if it does not already exist on disk.}
}
\value{
Saves a model to the provided directory for the user to load into the session using \code{limpiar_pos_load_model}.
}
\description{
A function that retrieves and downloads pre-built models made by the UDPipe community, covering 65 different languages. For more information on what models are available, visit:\href{https://bnosac.github.io/udpipe/docs/doc1.html#pre-trained-models}{UDPipe Documentation}
}
\examples{
model <- limpiar_pos_download_model(language = "english", model_directory = "data/", overwrite = TRUE)
}
