% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pos_import_model.R
\name{limpiar_pos_import_model}
\alias{limpiar_pos_import_model}
\title{Import UDPipe models to begin Parts of Speech Analysis}
\usage{
limpiar_pos_import_model(language)
}
\arguments{
\item{language}{The chosen language that the user wishes to select. There are 65 options to choose from}
}
\value{
Loads the model into memory, ready for the annotation steps in the parts of speech workflow
}
\description{
A function that retrieves and downloads pre-built models made by the UDPipe community, covering 65 different languages. For more information on what models are available, visit:\href{https://bnosac.github.io/udpipe/docs/doc1.html#pre-trained-models}{UDPipe Documentation}
}
\examples{
pos_model <- limpiar_pos_import_model(language = "english")
}
