% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pos_load_model.R
\name{limpiar_pos_load_model}
\alias{limpiar_pos_load_model}
\title{Load Previously Downloaded UDPipe Models}
\usage{
limpiar_pos_load_model(model_directory)
}
\arguments{
\item{model_directory}{A string pointing to a specific directory where a UDPipe classified model is saved}
}
\value{
Loads the model into memory, ready for the annotation steps in the parts of speech workflow
}
\description{
Load Previously Downloaded UDPipe Models
}
\examples{
model_loaded <- limpiar_pos_load_model(model_directory = "data/english-ewt-ud-2.5-191206.udpipe")
}
