% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pos_viz_cooccurrences.R
\name{limpiar_pos_viz_cooccurrences}
\alias{limpiar_pos_viz_cooccurrences}
\title{Visualize Parts of Speech Co-occurrences within sentences}
\usage{
limpiar_pos_viz_cooccurrences(
  annotated_df,
  chosen_pos = c("PROPN", "ADJ", "NOUN", "VERB", "ADV"),
  n_occurrences = 40
)
}
\arguments{
\item{annotated_df}{The output given from the limpiar_pos_annotate function as a data.frame object.}

\item{chosen_pos}{The chosen parts of speech the user wishes to visualise, tese can be either c("PROPN", "ADJ", "NOUN", "VERB").}

\item{n_occurrences}{How many different terms does the user wish to plot, the default is set to 40.}
}
\value{
A ggraph plot showing the connection between terms based on their occurrences within the same sentences.
}
\description{
Visualize Parts of Speech Co-occurrences within sentences
}
\examples{
limpiar_pos_viz_cooccurrences(annotated_df = output,
chosen_pos = c("PROPN", "ADJ"),
n_occurrences = 25)
}
