% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pp.R
\name{limpiar_pp_companies}
\alias{limpiar_pp_companies}
\title{Remove known companies for pits & peaks}
\usage{
limpiar_pp_companies(df, text_var)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{text_var}{Name of your text variable}
}
\value{
Data Frame or Tibble object with text variable edited inline
}
\description{
Remove known companies for pits & peaks
}
\examples{
\dontrun{
Example 1
df \%>\% limpiar_pp_companies(message)
Example 2
limpiar_pp_companies(df, message)
}
}
