% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pp.R
\name{limpiar_pp_products}
\alias{limpiar_pp_products}
\title{Replace entities for the Peaks&Pit classifier}
\usage{
limpiar_pp_products(df, text_var)
}
\arguments{
\item{df}{Data Frame or Tibble object}

\item{text_var}{Name of your text variable}
}
\value{
Data Frame or Tibble object with text variable edited inline
}
\description{
Replace entities for the Peaks&Pit classifier
}
