% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_retweets.R
\name{limpiar_retweets}
\alias{limpiar_retweets}
\title{limpiar_retweets}
\usage{
limpiar_retweets(df, x = .data$mention_content)
}
\arguments{
\item{df}{Name of the Data Frame or Tibble Object}

\item{x}{Name of the text variable/character vector}
}
\description{
Removes all posts with the 'rt' tag
}
\examples{
print("hello world")
}
