% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_shorthands.R
\name{limpiar_shorthands}
\alias{limpiar_shorthands}
\title{limpiar_shorthands}
\usage{
limpiar_shorthands(text_var, spaces_as_underscores = FALSE)
}
\arguments{
\item{text_var}{The text variable}

\item{spaces_as_underscores}{Whether multi-word corrections e.g. 'te quiero mucho' should have spaces or underscores. Default = FALSE}
}
\value{
The text variable with shorthands replaced
}
\description{
Replaces common Spanish shorthands and abbreviations with their longer form equivalents. Follows a similar design to limpiar_accents - so can be called inside a dplyr::mutate() function call
}
\examples{
\dontrun{
df <- df \%>\% mutate(text_var = limpiar_shorthands(text_var))
text_var <- limpiar_shorthands(text_var)
}
}
