% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_slang.R
\name{limpiar_slang}
\alias{limpiar_slang}
\title{limpiar_slang}
\usage{
limpiar_slang(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}
}
\value{
Data Frame or Tibble object with text variable altered
}
\description{
Replaces slang phrases from various Spanish dialects with everyday terms.
Function's primary use is to normalise text for Deep Learning sentiment algorithm.
Uses a hashing function and stringr::str_replace_all().
}
\examples{
\dontrun{
df \%>\%
limpiar_slang(text_var = text_var)}
}
