% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_spam_grams.R
\name{limpiar_spam_grams}
\alias{limpiar_spam_grams}
\title{Remove posts with suspicious n-grams}
\usage{
limpiar_spam_grams(
  data,
  text_var,
  n_gram = 8,
  top_n = 1000,
  min_freq = 5,
  in_parallel = TRUE
)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{text_var}{Name of the text variable}

\item{n_gram}{Number of words in the n-gram i.e. n = 2 = bigram}

\item{top_n}{Number of n-grams to keep}

\item{min_freq}{Minimum number of}

\item{in_parallel}{Whether to run the function with parallel processing}
}
\value{
A list with the suspicious-looking ngrams, removed posts, data & regex pattern
}
\description{
Function identifies posts which contain suspicious-looking n-gram patterns.
Posts can then be removed, the pattern inspected, and the posts that were removed too.
You can re-assign your current data frame to the 'clean' data frame through the third element of the list.
}
