% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_stopwords.R
\name{limpiar_stopwords}
\alias{limpiar_stopwords}
\title{limpiar_stopwords}
\usage{
limpiar_stopwords(text_var = text, stop_words)
}
\arguments{
\item{text_var}{name of the text variable}

\item{stop_words}{"sentiment" or "topics" - sentiment retains negtion cues\n}
}
\value{
the text variable with specified stopw ords removed
}
\description{
stop word list is editable via data("sentiment_stops") or data("topic_stops")
}
\examples{
\dontrun
df \%>\% mutate(text = limpiar_stopwords(text, "sentiment"))\%>\%
limpiar_spaces(text_var = text)
}
