% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_tags.R
\name{limpiar_tags}
\alias{limpiar_tags}
\title{Clean user handles and hashtags}
\usage{
limpiar_tags(df, text_var = mention_content, user = TRUE, hashtag = TRUE)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}

\item{user}{Whether to replace user handles or not TRUE = replace}

\item{hashtag}{Whether to replace hashtags or not TRUE = replace}
}
\value{
The Data Frame or Tibble object with user handles and/or hashtags removed from the text variable
}
\description{
Function replaces user handles and hashtags with neutral tags.
You can choose whether to replace both hashtags & users or either one.
}
\examples{
limpiar_examples

#Both user and hashtags
limpiar_examples \%>\% limpiar_tags() \%>\% dplyr::select(mention_content)

#Just user tags
limpiar_examples \%>\% limpiar_tags(hashtag = FALSE) \%>\% dplyr::select(mention_content)

#Just hashtags
limpiar_examples \%>\% limpiar_tags(user = FALSE) \%>\% dplyr::select(mention_content)

}
