% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_wrap_strings.R
\name{limpiar_wrap}
\alias{limpiar_wrap}
\title{Wrap strings for visual ease}
\usage{
limpiar_wrap(
  data,
  text_var = mention_content,
  n = 15,
  newline_char = "<br><br>"
)
}
\arguments{
\item{data}{Name of your Data Frame or Tibble object}

\item{text_var}{Name of your text variable. Can be given as a 'string' or a symbol - should refer to a column inside \code{data}}

\item{n}{number of words}

\item{newline_char}{the specific delimiter to wrap the texts with}
}
\value{
Data Frame with text variable edited in place
}
\description{
Useful for pre-processing a dataset in which you need to read many documents, or scan over a lot of documents, e.g. when rendering an interactive scatter plot and using plotly's hover, or when using \code{DT::datatable(escape = FALSE)}.
}
\examples{

limpiar_examples \%>\% limpiar_wrap(mention_content, n = 5, newline_char = "<br>")
}
