% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sentiment_stops}
\alias{sentiment_stops}
\title{sentiment_stops}
\format{
A tibble with 2 columns & 682 rows
\describe{
\item{Message}{Data frame with stop word vectors}
}
}
\usage{
data("sentiment_stops")
}
\description{
Stop words for sentiment analysis in Spanish (negation cues left in).
}
\keyword{internal}
