% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{text_var}
\alias{text_var}
\title{Helper for consistent documentation of \code{text_var}}
\arguments{
\item{text_var}{Name of your text variable. Can be given as a 'string' or a symbol - should refer to a column inside \code{data}}
}
\description{
Use \verb{@inheritParams text_var} to consistently document \code{text_var}.
}
\keyword{internal}
