% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_stops.R
\docType{data}
\name{topic_stops}
\alias{topic_stops}
\alias{sentiment_stops}
\title{topic_stops}
\format{
A tibble with 2 columns & 704 rows
\describe{
\item{Message}{Data frame with stop word vectors}
}
}
\usage{
data("topic_stops")
}
\description{
Stop words for topic modelling in Spanish.
}
\keyword{datasets}
