% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{code_browser_emojis}
\alias{code_browser_emojis}
\alias{spanish_emojis_df}
\title{spanish_emojis_df}
\format{
6 columns 1307 rows
\describe{
\item{Message}{Data frame with emoji conversions in Spanish}
}

3 columns 1853 rows
\describe{
\item{Message}{Data frame with emoji conversions}
}
}
\usage{
data("spanish_emojis_df")

data("code_browser_emojis")
}
\description{
data frame with columns for emoji converting in Spanish

data frame with columns for emoji converting
}
\keyword{internal}
