% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_browser_emojis.R, R/spanish_emojis_df.R
\docType{data}
\name{code_browser_emojis}
\alias{code_browser_emojis}
\title{code_browser_emojis}
\format{
3 columns 1853 rows
\describe{
\item{Message}{Data frame with emoji conversions}
}

6 columns 1307 rows
\describe{
\item{Message}{Data frame with emoji conversions in Spanish}
}
}
\usage{
data("code_browser_emojis")

data("spanish_emojis_df")
}
\description{
data frame with columns for emoji converting

data frame with columns for emoji converting in Spanish
}
\keyword{datasets}
