% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_accents.R
\name{limpiar_accents}
\alias{limpiar_accents}
\title{limpiar_accents}
\usage{
limpiar_accents(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}
}
\value{
text variable/character vector with accents replaced
}
\description{
Replaces accented characters with non-accented characters.
}
\examples{
\dontrun
df \%>\% limpiar_accents(text_var = message)

}
