% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_df.R
\name{limpiar_df}
\alias{limpiar_df}
\title{limpiar_df}
\usage{
limpiar_df(df, text_var)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{text_var}{The text variable/character vector}
}
\value{
The original data frame with the text variable cleaned
}
\description{
limpiar_df
}
\examples{
print("hello world")
}
