% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_df.R
\name{limpiar_df}
\alias{limpiar_df}
\title{Call multiple LimpiaR functions together}
\usage{
limpiar_df(
  df,
  text_var = mention_content,
  remove_retweets = TRUE,
  remove_punctuation = FALSE
)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{text_var}{The text variable/character vector}

\item{remove_retweets}{Whether or not to remove retweets, default is to remove}

\item{remove_punctuation}{Whether or not to remove punctuation, default is not to remove}
}
\value{
The original Data Frame with the text variable cleaned
}
\description{
Used on a data frame to remove duplicates, empty rows, and retweets.
Transform the text variable to lower case and remove user handles and hashtags.
Clean accents, urls, shorthands, repeated characters, and if specified punctuation.
}
\details{
You should only use this function when you are sure you want to execute each of the steps.
}
\examples{
limpiar_df(limpiar_examples, text_var = mention_content, remove_punctuation = TRUE)

}
