% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_df.R
\name{limpiar_df}
\alias{limpiar_df}
\title{Call multiple LimpiaR functions together}
\usage{
limpiar_df(
  df,
  text_var = mention_content,
  remove_retweets = TRUE,
  remove_punctuation = FALSE
)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{text_var}{The text variable/character vector}

\item{remove_retweets}{Whether or not to remove retweets, default is to remove}

\item{remove_punctuation}{Whether or not to remove punctuation (and then clean whitespace)}
}
\value{
The original Data Frame with the text variable cleaned
}
\description{
used on a data frame to remove duplicates, empty rows, and retweets,
turn the text variable to all lower case (advisable to do any part of speech tagging before using).
Function also cleans accents, common urls, punctuation if specified and trims excess white spaces.
Use remove_retweets = FALSE to keep retweets in the text variable.
}
\examples{
\dontrun{limpiar_df(df, text_var = mention_content)}

}
