% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_emojis.R
\name{limpiar_emojis}
\alias{limpiar_emojis}
\title{limpiar_emojis}
\usage{
limpiar_emojis(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of Data Frame or Tibble Object}

\item{text_var}{Name of text variable}
}
\value{
Data Frame with most emojis cleaned
}
\description{
functions to replace most emojis with a snakecase descrition + '_emoji'
}
\examples{
\dontrun{df \%>\% limpiar_emojis(text_var)}
}
