% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_inspect.R
\name{limpiar_inspect}
\alias{limpiar_inspect}
\title{limpiar_inspect}
\usage{
limpiar_inspect(
  data,
  pattern,
  text_var = mention_content,
  post_url = mention_url,
  title = "inspect"
)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{pattern}{Pattern you wish to inspect e.g. "link bio"}

\item{text_var}{Name of the text variable/character vector}

\item{post_url}{Name of the data frame's URL-column}

\item{title}{Name of the viewable pane}
}
\description{
Produces a viewable data frame with posts matching a regular expression and their url. Useful for investigating suspected spam posts.
}
\examples{
df <- data.frame(text_variable = rbind("check me out", "don't look at me"),text_url = rbind("www.twitter.com", "www.facebook.com"))
limpiar_inspect(df, "check", text_var = text_variable, post_url = text_url)
}
