% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_inspect.R
\name{limpiar_inspect}
\alias{limpiar_inspect}
\title{limpiar_inspect}
\usage{
limpiar_inspect(
  data,
  pattern,
  text_var = .data$mention_content,
  post_url = .data$mention_url
)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{pattern}{Pattern you wish to inspect e.g. "link bio"}

\item{text_var}{Name of the text variable/character vector}

\item{post_url}{Name of the data frame's URL-column}
}
\description{
Produces a viewable data frame with posts matching a regular expression and their url. Useful for investigating suspected spam posts.
}
\examples{
print("hello world")
}
