% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_link_click.R
\name{limpiar_link_click}
\alias{limpiar_link_click}
\title{Prepare a URL column to be clickable in Shiny/Data Table}
\usage{
limpiar_link_click(df, url_var)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{url_var}{URL Column}
}
\value{
data frame with URL column edited to be clickable
}
\description{
Will allow you to click the hyperlink to load a URL, e.g. for selecting an image.
}
\details{
Make sure that DataTable is rendered with the argument 'escape = FALSE' or column will be all text.

The function now checks if your url_var was a clickable link, and if it is then it won't add any new formatting.
}
\examples{
df <- LimpiaR::limpiar_examples[1, ]
df["mention_url"]

df <- df \%>\% limpiar_link_click(mention_url)
df["mention_url"]

}
