% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_na_cols.R
\name{limpiar_na_cols}
\alias{limpiar_na_cols}
\title{limpiar_na_cols}
\usage{
limpiar_na_cols(data, y)
}
\arguments{
\item{data}{The Data Frame or Tibble Object}

\item{y}{A decimal between 0 and 1: The minimum proportion of NA values a column should contain to not be deleted.}
}
\value{
data frame or tibble
}
\description{
A function to remove columns with a specified proportion of NA/missing values.
}
\examples{
print("hello world")
}
