% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_na_cols.R
\name{limpiar_na_cols}
\alias{limpiar_na_cols}
\title{limpiar_na_cols}
\usage{
limpiar_na_cols(df, threshold)
}
\arguments{
\item{df}{The Data Frame or Tibble object}

\item{threshold}{Threshold of non-NA entries a column must exceed to be retained.}
}
\value{
Data Frame or Tibble with NA-heavy columns purged
}
\description{
A function to remove NA-heavy columns. Retains columns
}
\examples{
a <- cbind(c(NA, NA, NA, NA, NA))
b <- cbind(c(NA, NA, 2, NA, 3))
c <- cbind(c(1, 2, 3, 4, 5))
df <- tibble::as_tibble(data.frame(a, b, c))
df <- limpiar_na_cols(df, 0.4)

}
