% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_pos_annotate.R
\name{limpiar_pos_annotate}
\alias{limpiar_pos_annotate}
\title{Annotate Texts for Parts of Speech Analysis}
\usage{
limpiar_pos_annotate(
  data,
  text_var,
  saved_model,
  in_parallel = FALSE,
  parse_text = FALSE,
  ...
)
}
\arguments{
\item{data}{The data.frame or tibble object containing any texts that the user wishes to conduct parts of speech analysis on.}

\item{text_var}{Any texts or sentences the user wishes to perform the parts of speech annotations on.}

\item{saved_model}{A UDPipe model loaded into memory using \code{limpiar_pos_load_model}, and must be of class 'udpipe_model'.}

\item{in_parallel}{A logical argument allowing the user to leverage parallel processing to speed the function up. If set to TRUE,  the function will select the number of available cores minus 1 and process more efficiently(faster).}

\item{parse_text}{Whether to perform dependency parsing on tokens. The default is set to "parser". The user may skip this step by supplying the argument with "none".}

\item{...}{To enable the user to supply any additonal arguments.}
}
\value{
Returns a data frame object derived from CONLL-U formatting, where each row is an annotation of a word. To find out more on the formatting methods, read \href{https://universaldependencies.org/format.html}{here}.
}
\description{
Annotate Texts for Parts of Speech Analysis
}
\examples{
output <- limpiar_pos_annotate(data = data, text_var = clean_text, saved_model = model_loaded, in_parallel = TRUE, parser = "none")
}
