% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_retweets.R
\name{limpiar_retweets}
\alias{limpiar_retweets}
\title{Clean retweets from the text variable}
\usage{
limpiar_retweets(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of the Data Frame or Tibble Object}

\item{text_var}{Name of the text variable/character vector}
}
\description{
Removes all posts with the 'rt' or 'RT' tag. Useful when visualising organic posts.
Particularly useful in conjunction with the ParseR & SegmentR visualisations.
}
\examples{
df <- data.frame(text_variable = cbind(c("rt <3", "RT <3", "original tweet")))
limpiar_retweets(df, text_variable)
}
