% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_shorthands.R
\name{limpiar_shorthands}
\alias{limpiar_shorthands}
\title{limpiar_shorthands}
\usage{
limpiar_shorthands(
  df,
  text_var = mention_content,
  spaces_as_underscores = FALSE
)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}

\item{spaces_as_underscores}{Whether multi-word corrections e.g. 'te quiero mucho' should have spaces or underscores. Default = FALSE}
}
\value{
The text variable with shorthands replaced
}
\description{
Replaces common Spanish shorthands and abbreviations with their longer form equivalents.
}
\examples{
\dontrun{
df <- df \%>\% mutate(text_var = limpiar_shorthands(text_var))
text_var <- limpiar_shorthands(text_var)
}
}
