% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_slang.R
\name{limpiar_slang}
\alias{limpiar_slang}
\title{Clean slang from multiple Spanish dialects}
\usage{
limpiar_slang(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}
}
\value{
Data Frame or Tibble object with text variable altered
}
\description{
Replaces slang phrases from various Spanish dialects with everyday terms.
Function's primary use is to normalise text for Deep Learning sentiment algorithm.
Care should be taken when using this function, e.g. panda -> grupo, as that is by far the most common usage in the texts we use.
However, in a data set where many people talk about panda bears or 'oso panda', there will be unwanted changes.
I have tried to avoid this problem where possible, by including things like 'me la suda' instead of changing 'suda'.
}
\examples{
\dontrun{
df \%>\%
limpiar_slang(text_var = text_var)}
}
