% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_spaces.R
\name{limpiar_spaces}
\alias{limpiar_spaces}
\title{limpiar_spaces}
\usage{
limpiar_spaces(df, text_var = mention_content)
}
\arguments{
\item{df}{Name of the Data Frame or Tibble object}

\item{text_var}{Name of the text variable/character vector}
}
\value{
text variable/character vector with excess spaces removed
}
\description{
Remove excess spaces from the text variable
}
\examples{
df <- data.frame(text_variable = "clean   the   spaces please")
limpiar_spaces(df, text_var = text_variable)
}
