% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_spaces.R
\name{limpiar_spaces}
\alias{limpiar_spaces}
\title{limpiar_spaces}
\usage{
limpiar_spaces(df, x = .data$mention_content)
}
\arguments{
\item{df}{Name of the Data Frame or Tibble object}

\item{x}{Name of the text variable/character vector}
}
\value{
text variable/character vector with excess spaces removed
}
\description{
Remove excess spaces from the text variable
}
\examples{
print("hello world")
}
