% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_stopwords.R
\name{limpiar_stopwords}
\alias{limpiar_stopwords}
\title{limpiar_stopwords}
\usage{
limpiar_stopwords(df, text_var = mention_content, stop_words)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{name of the text variable}

\item{stop_words}{"sentiment" or "topics" - sentiment retains negtion cues\n}
}
\value{
the text variable with specified stopw ords removed
}
\description{
stop word list is editable via data("sentiment_stops") or data("topic_stops")
}
\examples{
\dontrun
df \%>\% limpiar_stopwords(stop_words = "sentiment") \%>\%
limpiar_spaces(text_var = text)
}
