% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_stopwords.R
\name{limpiar_stopwords}
\alias{limpiar_stopwords}
\title{Clean stop words for visualisations}
\usage{
limpiar_stopwords(df, text_var = mention_content, stop_words)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{name of the text variable}

\item{stop_words}{"sentiment" or "topics" - sentiment retains negation cues}
}
\value{
the text variable with stop words from specified list removed
}
\description{
The two lists - sentiment & topics, are very similar, in that most words are in both lists.
However, sentiment analysis is sensitive to negation, so negation cues e.g.
"no", "nada" etc. are not removed by the sentiment list. For most purposes, topics
are the go-to lists, but care is always advised when removing stop words.
}
\details{
stop word list is editable via data("sentiment_stops") or data("topic_stops").
}
\examples{
limpiar_examples \%>\% dplyr::select(mention_content)

limpiar_examples \%>\% limpiar_stopwords(stop_words = "topics") \%>\%
dplyr::select(mention_content) \%>\% limpiar_spaces()

}
