% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_tags.R
\name{limpiar_tags}
\alias{limpiar_tags}
\title{limpiar_tags}
\usage{
limpiar_tags(df, text_var = mention_content, user = TRUE, hashtag = TRUE)
}
\arguments{
\item{df}{Name of Data Frame or Tibble object}

\item{text_var}{Name of text variable/character vector}

\item{user}{Whether to replace user handles or not TRUE = replace}

\item{hashtag}{Whether to replace hashtags or not TRUE = replace}
}
\value{
the df object with the text_var cleaned of user / # tags
}
\description{
Function replaces user handles and hashtags and replaces them with neutral tags.
You can choose whether to replace both hashtags & users or either one.
}
