% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_url.R
\name{limpiar_url}
\alias{limpiar_url}
\title{Clean URLs from the text variable}
\usage{
limpiar_url(df, text_var = mention_content)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{text_var}{Name of the text variable/character vector}
}
\value{
Data Frame or Tibble object with URLs removed from the text variable
}
\description{
Removes the most common forms of URLs from the text variable
}
\examples{
limpiar_examples \%>\% dplyr::select(mention_content)

limpiar_examples \%>\% limpiar_url() \%>\% dplyr::select(mention_content)

}
