% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_url.R
\name{limpiar_url}
\alias{limpiar_url}
\title{limpiar_url}
\usage{
limpiar_url(df, text_var = .data$mention_content)
}
\arguments{
\item{df}{Data Frame or Tibble Object}

\item{text_var}{Name of the text variable/character vector}
}
\value{
text variable/character vector with URLs removed
}
\description{
Removes the most common forms of URLs from a text variable of a data frame/tibble
}
\examples{
\dontrun
df \%>\% limpiar_url()
}
